+++ src/core/affinity.h
@@ -3,11 +3,11 @@
 
 #include "parseobj.h"
 
-void move(model *m, double x, double y, double z);
-void _move(float *array, size_t size, double x, double y, double z);
-void rotate(model  *m, double angle, char axis);
-void _rotate(float *array, double angle, size_t size, char axis);
-void scale(model *m, double xs, double ys, double zs);
-void _scale(float *array, size_t size, double xs, double ys, double zs);
+void move(model *m, float x, float y, float z);
+void _move(float *array, size_t size, float x, float y, float z);
+void rotate(model  *m, float angle, char axis);
+void _rotate(float *array, float angle, size_t size, char axis);
+void scale(model *m, float xs, float ys, float zs);
+void _scale(float *array, size_t size, float xs, float ys, float zs);
 
 #endif /*  _SRC_AFFINITY_H */
